airflow_resources/requirements-airflow.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
unidecode==1.3.7
apache-airflow-providers-docker==3.8.1
